requirements.txt
numpy
tensorflow==1.13.1
keras==2.1.6
pillow
keras2onnx==1.17.0
onnxruntime
h5py==2.10.0
